renaming of directories in order to use more consistent names throughout the library
