+++ include/exadg/convection_diffusion/user_interface/input_parameters.h
@@ -29,7 +29,7 @@
 #include <exadg/convection_diffusion/user_interface/enum_types.h>
 #include <exadg/grid/enum_types.h>
 #include <exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.h>
-#include <exadg/solvers_and_preconditioners/preconditioner/enum_types.h>
+#include <exadg/solvers_and_preconditioners/preconditioners/enum_types.h>
 #include <exadg/solvers_and_preconditioners/solvers/enum_types.h>
 #include <exadg/solvers_and_preconditioners/solvers/solver_data.h>
 #include <exadg/time_integration/enum_types.h>
