+++ include/exadg/incompressible_navier_stokes/spatial_discretization/operator_projection_methods.cpp
@@ -20,8 +20,8 @@
  */
 
 #include <exadg/incompressible_navier_stokes/spatial_discretization/operator_projection_methods.h>
-#include <exadg/poisson/preconditioner/multigrid_preconditioner.h>
-#include <exadg/solvers_and_preconditioners/preconditioner/jacobi_preconditioner.h>
+#include <exadg/poisson/preconditioners/multigrid_preconditioner.h>
+#include <exadg/solvers_and_preconditioners/preconditioners/jacobi_preconditioner.h>
 #include <exadg/solvers_and_preconditioners/util/check_multigrid.h>
 
 namespace ExaDG
