+++ include/exadg/incompressible_navier_stokes/user_interface/input_parameters.h
@@ -30,7 +30,7 @@
 #include <exadg/incompressible_navier_stokes/user_interface/enum_types.h>
 #include <exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.h>
 #include <exadg/solvers_and_preconditioners/newton/newton_solver_data.h>
-#include <exadg/solvers_and_preconditioners/preconditioner/enum_types.h>
+#include <exadg/solvers_and_preconditioners/preconditioners/enum_types.h>
 #include <exadg/solvers_and_preconditioners/solvers/solver_data.h>
 #include <exadg/time_integration/enum_types.h>
 #include <exadg/time_integration/restart_data.h>
