+++ include/exadg/solvers_and_preconditioners/multigrid/multigrid_preconditioner_base.h
@@ -34,7 +34,7 @@
 #include <exadg/solvers_and_preconditioners/multigrid/multigrid_input_parameters.h>
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/smoother_base.h>
 #include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer.h>
-#include <exadg/solvers_and_preconditioners/preconditioner/preconditioner_base.h>
+#include <exadg/solvers_and_preconditioners/preconditioners/preconditioner_base.h>
 
 // forward declarations
 namespace ExaDG
